tests: Move bulk of tests into base class in preparation for MP3 tests
