+++ tests/Flac/FlacTaggingTest.php
@@ -2,20 +2,18 @@
 
 namespace IndieHD\AudioManipulator\Tests\Flac;
 
-use PHPUnit\Framework\TestCase;
+use IndieHD\AudioManipulator\Tests\Tagging\TaggingTest;
 
 use Symfony\Component\Filesystem\Exception\FileNotFoundException;
 
 use IndieHD\AudioManipulator\Processing\ProcessFailedException;
 use IndieHD\AudioManipulator\Tagging\AudioTaggerException;
 
-class FlacTaggingTest extends TestCase
+class FlacTaggingTest extends TaggingTest
 {
     private $testDir;
 
     private $tmpDir;
-
-    private $sampleDir;
 
     private $sampleFile;
 
@@ -26,6 +24,8 @@ class FlacTaggingTest extends TestCase
      */
     public function setUp(): void
     {
+        $this->setFileType('flac');
+
         $this->testDir = __DIR__ . DIRECTORY_SEPARATOR . '..'
             . DIRECTORY_SEPARATOR;
 
@@ -46,192 +46,8 @@ class FlacTaggingTest extends TestCase
             ->create($this->tmpFile);
     }
 
-    /**
-     * Ensure that a FLAC file can be tagged with metadata.
-     *
-     * @return void
-     */
-    public function testItCanTagFlacFile()
+    protected function setFileType(string $type): void
     {
-        $tagData = [
-            'title' => ['Test Song'],
-            'artist' => ['Foobius Barius'],
-            'date' => ['1981'],
-            'description' => ['All rights reserved.'],
-            'album' => ['Test Title'],
-            'discnumber' => ['1'],
-            'tracknumber' => ['1'],
-            'genre' => ['Rock'],
-        ];
-
-        $this->flacManipulator->writeTags(
-            $tagData
-        );
-
-        $fileDetails = $this->flacManipulator
-            ->tagger
-            ->getid3
-            ->analyze($this->flacManipulator->getFile());
-
-        $this->assertEquals(
-            [
-                'title' => $tagData['title'],
-                'artist' => $tagData['artist'],
-                'date' => $tagData['date'],
-                'description' => $tagData['description'],
-                'album' => $tagData['album'],
-                'discnumber' => ['1'],
-                'tracknumber' => [$tagData['tracknumber'][0]],
-                'genre' => ['Rock'],
-            ],
-            $fileDetails['tags']['vorbiscomment']
-        );
-    }
-
-    public function testItThrowsExceptionWhenFileDoesNotExist()
-    {
-        $this->expectException(FileNotFoundException::class);
-
-        $this->flacManipulator = $this->flacManipulatorCreator
-            ->create('foo.bar');
-
-        $this->flacManipulator->writeTags(
-            []
-        );
-    }
-
-    public function testItThrowsExceptionWhenProcessFails()
-    {
-        $this->expectException(ProcessFailedException::class);
-
-        $this->flacManipulator->tagger->command->setBinary('non-existent-binary-path');
-
-        $this->flacManipulator->writeTags([]);
-    }
-
-    private function removeAllTags()
-    {
-        $this->flacManipulator->tagger->removeAllTags(
-            $this->flacManipulator->getFile()
-        );
-    }
-
-    private function removeArtwork()
-    {
-        $this->flacManipulator->tagger->removeArtwork(
-            $this->flacManipulator->getFile()
-        );
-    }
-
-    private function embedArtwork()
-    {
-        $this->removeArtwork();
-
-        $this->flacManipulator->tagger->writeArtwork(
-            $this->flacManipulator->getFile(),
-            $this->sampleDir . 'flac-logo.gif'
-        );
-    }
-
-    public function testItCanEmbedArtwork()
-    {
-        $this->embedArtwork();
-
-        $fileDetails = $this->flacManipulator->tagger->getid3->analyze(
-            $this->flacManipulator->getFile()
-        );
-
-        $testImage = file_get_contents($this->sampleDir . 'flac-logo.gif');
-
-        $this->assertEquals(
-            $testImage,
-            $fileDetails['comments']['picture'][0]['data']
-        );
-
-        $this->assertEquals(
-            $testImage,
-            $fileDetails['flac']['PICTURE'][0]['data']
-        );
-    }
-
-    public function testItCanRemoveArtworkFromFlacFile()
-    {
-        $this->removeAllTags();
-
-        $this->embedArtwork();
-
-        $this->removeArtwork();
-
-        $fileDetails = $this->flacManipulator->tagger->getid3->analyze(
-            $this->flacManipulator->getFile()
-        );
-
-        $this->assertArrayNotHasKey('comments', $fileDetails);
-
-        $this->assertArrayNotHasKey('PICTURE', $fileDetails['flac']);
-    }
-
-    public function testItCanRemoveTagsFromFlacFile()
-    {
-        $this->removeAllTags();
-
-        $this->flacManipulator->writeTags(
-            [
-                'title' => ['Foo'],
-            ]
-        );
-
-        $this->flacManipulator->removeTags(
-            [
-                'title',
-            ]
-        );
-
-        $fileDetails = $this->flacManipulator->tagger->getid3->analyze(
-            $this->flacManipulator->getFile()
-        );
-
-        $this->assertArrayNotHasKey('tags', $fileDetails);
-    }
-
-    public function testItCanWriteUtf8TagValuesAccurately()
-    {
-        $tagData = [
-            'title' => ['﻿ᚠᛇᚻ᛫ᛒᛦᚦ᛫ᚠᚱᚩᚠᚢᚱ᛫ᚠᛁᚱᚪ᛫ᚷᛖᚻᚹᛦᛚᚳᚢᛗ'],
-        ];
-
-        $this->flacManipulator->writeTags(
-            $tagData
-        );
-
-        $fileDetails = $this->flacManipulator
-            ->tagger
-            ->getid3
-            ->analyze($this->flacManipulator->getFile());
-
-        $this->assertEquals(
-            [
-                'title' => $tagData['title'],
-            ],
-            $fileDetails['tags']['vorbiscomment']
-        );
-    }
-
-    public function testExceptionIsThrownWhenTagsCannotBeVerifiedAfterWriting()
-    {
-        $this->expectException(AudioTaggerException::class);
-
-        // Set an inappropriate encoding, which will cause the tag to be
-        // written incorrectly.
-
-        $this->flacManipulator->tagger->setEnv(['LC_ALL' => 'en_US.iso-8859-1']);
-
-        $tagData = [
-            'title' => ['﻿ᚠᛇᚻ᛫ᛒᛦᚦ᛫ᚠᚱᚩᚠᚢᚱ᛫ᚠᛁᚱᚪ᛫ᚷᛖᚻᚹᛦᛚᚳᚢᛗ'],
-        ];
-
-        $this->flacManipulator->writeTags(
-            $tagData
-        );
+        $this->fileType = $type;
     }
 }
